Update to allow for install into EECMS 2.9.0
More or less a hot fix or hack to bypass error on installation in 2.9.0

Installation in 2.9.0 was throwing errors on installation because of missing default settings in exp_extension table. Added line 29 and 42 to add do the activation array. Not a solid fix (due to the inability to edit settings), but works through testing on 2.9.0.